fastapi-app/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
fastapi==0.63.0
gunicorn==20.0.4
Pillow==7.0.0
pydantic==1.7.3
starlette==0.13.6
streamlit==0.70.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
uvicorn==0.13.3